Extract context card lifestyle editors
