+++ service-worker.js
@@ -104,6 +104,7 @@ const APP_SHELL = [
   '/js/context-card-summaries.js',
   '/js/context-card-editor-ui.js',
   '/js/context-card-medical-history-editor.js',
+  '/js/context-card-lifestyle-editors.js',
   '/js/focus-card.js',
   '/js/onboarding-view.js',
   '/js/emf-facade.js',
